Move `stem()` to separate file
