+++ src/sdr/plot/__init__.py
@@ -4,8 +4,8 @@ A subpackage for various plotting functions using :obj:`matplotlib`.
 
 from ._detection import *
 from ._filter import *
-from ._helper import *
 from ._modulation import *
 from ._rc_params import *
 from ._spectral_estimation import *
 from ._time_domain import *
+from ._utility import *
